moves type delarration to generic types
